requirements.txt
opencv-python==4.7.0.72
Pillow==9.5.0
pylibdmtx==0.1.10